// ui/src/auth/types.ts
export type UserType = Record<string, any> | null;

export type AuthState = {
  user: UserType;
  loading: boolean;
};

export type AuthContextValue = {
  user: UserType;
  loading: boolean;
  authenticated: boolean;
  unauthenticated: boolean;
  superAdminRole?: number | null;
  superAdminLoading?: boolean;
  checkUserSession?: () => Promise<void>;
};